+++ app/models/server_model.py
@@ -23,3 +23,16 @@ class Server:
         if result is not None:
             return cls(nombre_servidor = result)
         return None
+    
+    @classmethod
+    def get_serverUsuario(cls,usuario):                  #no está implementado
+        query="""SELECT s.nombre_servidor FROM servidor s
+                 JOIN usuario_servidor us ON s.id_servidor = us.servidor
+                 JOIN usuarios u ON u.id_usuario = us.usuario
+                 WHERE u.correo = %(correo)s;"""
+        params = usuario.__dict__
+        result = DatabaseConnection.fetch_all(query,params=params)
+        
+        if result is not None:
+            return cls(nombre_servidor = result)
+        return None
