+++ app/routes/user_bp.py
@@ -6,5 +6,5 @@ user_bp = Blueprint('user_bp', __name__)
 
 user_bp.route('/alias', methods=['GET'])(UsuarioController.getAlias)
 user_bp.route('/cargar_servidores', methods=['GET'])(ServerController.getServers)
-# user_bp.route('/servidores_del_usuario', methods=['GET'])(ServerController.getServersUsuario)
+user_bp.route('/servidores_del_usuario', methods=['GET'])(ServerController.getServersUsuario)
 user_bp.route('/register', methods=['GET','POST'])(UsuarioController.register)
